Let .uo and .ui files have $(envvar) prefixes in their paths
This will let these files specify positions for source code that is
independent of an absolute position in the file-tree hierarchy, but is
relative to a fixed base. The first step is to allow $(HOLDIR) so that
HOL can have its own .uo files be easily relocatable.

Next step is to write .uo and .ui files using this facility.