Complete relocation-build feature
Closes #105

To be explicit, after building HOL in one directory one can then

$ bin/build cleanForReloc
$ cd ..
$ mv HOL somewhere_else
$ cd somewhere_else
$ poly < tools/smart-configure.sml
$ bin/build --relocbuild

The last step (the "relocation" build) takes less than a minute and
results in a functional system that is housed in a different location.

Next step on this branch of work is to provide tools allowing other
developments (those not under HOLDIR) to be moved in a similar way.