hol-mode: add hol-print-info
I recently have the problem on various machines, that have multiple
copies of HOL installed, to figure out which version is started by hol-
mode. I did this by looking at the value of HOLDIR, the current load
paths etc in ML and the used hol and holmake executables in Lisp. To
simplify this process of looking up these values, this commit adds this
functionality to hol-mode.