Use standard (>>, >-, >|) symbolic connectives for THEN*
This is done under src/, replacing Joe Hurd's earlier versions of the
same which were ++, >> and << respectively. His proofs also
occasionally rely on the fact that he made >> and >|
right-associative, which is a great idea. His symbols are inherently
bad but they clash with existing practice, and there seems to be a
good argument to trying to keep material in src reasonably
consistent. (Using >> for THEN1 instead of THEN is particularly
confusing given our current practice.)