Fix Holmake's -I flag so it is not applied recursively
The -I flag to Holmake was being re-interpreted with each recursive
entry of a fresh directory, so that when the included directory was
entered, the Holmake process then attempted to again include that path
in the newly entered directory. This is fine if the path is
absolute (the loop detection will stop the recursion), but if it's
relative, the path won't be present in that directory and a
file-system IO error occurs.

Closes #413