Allow Q.MATCH_RENAME_TAC (and friends) to backtrack over pat. parses
When a pattern is passed to `Q.MATCH_RENAME_TAC` (and others in this
family), it is parsed once (making a fixed choice for overloaded
names) and then used as a matching-pattern.  The behaviour should be
that all possible overload-choices are used as possible patterns.

Less clear is how the choices in the assumption and subterm based
tactics should be ordered: for example it might be:

  <choose parse> -> <choose assumption/subterm for match>

The current implementation has subterm selection happening last, but
assumption selection happening first.  So, the ordering for
MATCH_ASMSUB_RENAME_TAC is:

   find an assumption
      find a parse
         find a sub-term

Regardless of this order, the parsing only happens once so this
shouldn't any difference to efficiency. This ordering can make a
difference to ultimate user-visible behaviour.

Includes some test-cases.

Closes #430