Remove complicated behaviour behind Unicode trace/flag
Now the Unicode trace is just an inverted handle on the behaviour behind
the PP.avoid_unicode trace. I couldn't use the alias machinery for
this because of the inversion, but they can still alias to the same
underlying SML reference.

While the ProvideUnicode structure still exists it is dramatically
simpler (I decided I didn't want to cut and paste it into the already-
quite-long-enough Parse.sml).

Closes #363