Move absyn post-processing done for let-expressions to boolpp
This removes this behaviour from the core-parser, which is a win for
simplicity, and makes it use the same (standard) API as other
absyn-postprocessors (such as those used to implement monad syntax).

This modularisation is a step towards making the let-syntax richer (as
per github issue #154, which calls for Isabelle-style syntax).