Use Unicode quotes (“...”) rather than --` ... `-- and ``...``
Need to do this in boolScript felt particularly pressing given the --
syntax. This is a horrible holdover, and consumes many columns
(particularly when the frequently required parentheses are also
included).

I think there are relatively simple regular expressions for doing this
automatically.  For the moment, bool and arithmetic are two
trailblazers.