Put another small example into Travis self-test sequence
This required a change in the relevant Holmakefile so that the
hol88Lib.sml file got seen and included appropriately when not part of
the standard build sequence (Travis builds use custom build sequences
so that the tests stand a chance of completing in good time).

This particular test chosen because it would have made the error fixed
in 21392d97b apparent from Travis testing.  I hope there are other
interesting bits of code exercised by this same example.