Fix some infelicities in holyhammer build details
Trying to build with Moscow ML revealed a dependency on code in
tactictoe, so I reordered the directories in the build sequence. But
then, tactictoe is only built under Poly/ML anyway.  So, I added the
[poly] annotation to the holyhammer directories too.

Changing the order may then seem unnecessary, but if one was to try to
exercise any of the holyhammer code, as part of a regression test for
example, then the tactictoe code does need to have been processed and
put in sigobj for that to work.