Add finite solvability example
Formalises (an improvement of) the main result in:
"Solving Problems with Finite Test Sets"
  by Cristian S. Calude, Helmut Jürgensen, Shane Legg

Every first-order formula over the natural numbers is "solved" by a
single instance. That is, there is an assignment of variables such that
the original formula is true iff it is true under that assignment
(ignoring all the quantifiers).