Fix bug in prove_case_elim_thm
Test-case tweaks old bug, which was caused by conflicting names. If
the first constructor in the type was

  C ty

and a later constructor was D (ty -> bool), then the code would end up
with a theorem of the form

   x = D f /\ ... |- caseconst x f f1 f2

where the f in the hypotheses corresponded to the argument to the D,
and the f in the conclusion corresponded to the function passed to the
case-constant, taking a ty and returning (in this case) a bool.  Then,
when CHOOSE tried to wrap an existential around the f in the
hypothesis, it would fail because we'd inadvertently identified two
variables that were supposed to be different.

Fix is to be even more paranoid about choice of variable names when
stripping the existentials arising in type's nchotomy theorem.

Bug was apparent in examples/lambda/typing/type_schemasScript.sml