Do "last made by" checks even when cleaning in Holmake
This reverts 550b90aca, which doesn't come with enough justification
for me to understand the problem that I thought I was solving in 2010.
The problem with omitting the checks is that if the "wrong" Holmake is
called, and if the Holmakefile contains references to things like
$(HOLDIR), then the wrong cleaning action can occur.

For example, if an INCLUDES line refers to $(HOLDIR), then a recursive
clean will recurse into the wrong directory.  Similarly, if an
EXTRA_CLEANS line refers to $(HOLDIR), then the wrong file will get
deleted.