Rationalise handling of Unicode lowercase phi characters
The "mainline" phi character in Unicode is U+03C6, which in most fonts
looks more like what TeX calls a \varphi. In the font you are using to
view this message it is "φ". There is also an additional U+03D5, which
is written the way Knuth wanted his \phi to look (in this font: "ϕ").
If these are seen in HOL input, EmitTeX now maps them appropriately.
Similarly, the Emacs mode for producing them now maps C-S-p h to the
second and C-S-p v to the first.