Tweak printing of long function applications
This brings the behaviour (on at least the examples I've tested) back
to the way things appeared in Kananaskis-11.

The change in the munger test only consists of some extra space
characters appearing in the output on a line by
themselves.  (Previously the line had just consisted of a newline
without the preceding spaces.)