EmitTeX: emit math-mode "spaces" after turnstiles properly
Again, thanks to Ramana for the bug report. Document the fact that the
math-mode will probably require overriding $Turnstile$ to get best
results. With the default \;\; spacing, results without an override
are passable, but I'm not so keen on \;\; spacing these days.

With a \; spacing and an 11pt font (Computer Modern in my
experiments), the appropriate overload seems to be

$Turnstile$ 6 \ensuremath{\HOLTokenTurnstile{}\;}

The \ensuremath is there for if and when the same overrides file is
used in a "nomath" setting.