first attempt at porting tamarack2 example
This commits adds the files just previously mentioned in the readme to
examples/hardware. These are hol88 examples that should be ported to the
current version of HOL 4. I started with porting the tamarack2 example.
A very close port running on HOL 4 is ready, a more liberal one is work-
in-progress.