Fix pretty-printing of list-forms
Issue was that pretty-printing blocks need to be around the list
delimiters (and included elements), not around just the elements with
the delimiters outermost.  The block_info field of the listspec
constructor is now ignored when these are printed; instead the binfo
is lifted (by term_grammar.add_listform) to the rule in which the
listspec appears.

Closes #529