Prove a course of values recursion combinator for prim. recursion
WFM makes it easier to define functions that recurse on arguments that
are arbitrarily smaller than the input argument (rather than just one
smaller), embodying the construction that creates an encoded list of
all possible answers.

This machinery is not as general as it might be, because it only
allows for unary functions. To be better, you'd want to allow for
other arguments, and (moreover) these should be able to vary. Given
the complicated definition of pr_nsub (which I don't now really
understand) in computability/lambda/prterm, I think this might be
quite involved. (There are resources available on the web describing
the general construction though.)