Allow custom p/printers to fire more reliably when in "rator" pos'n
The treatment of COND worked because of special-handling for "case"
constants, and for LET it works because the special printer just
checks the underlying head constant. However, for trickier patterns,
you need to call Overload.strip_comb, and if this succeeds for the
larger term, it needs to succeed when called again on the rator
sub-term.  Previously, for no obvious reason, the second call would
effectively report that there was no pattern to be had.

Add test-case showing desired behaviour with monad pretty-printer.

Closes #522