Lex tokens consisting of only $-signs as identifiers
This adds behaviour to the system: previously it was impossible to get
anything out of quotations containing bare $-signs. Now these are
lexed as term identifiers. (The type parser was already capable of
lexing such things sensibly.)

I am strongly inclined to now use single dollar-signs as in Haskell,
as an alternative form of adding parentheses. For example,

   f $ g $ x + 1

is a lighter way of writing f (g (x + 1)).  The effect is achieved by
making $ a right-associative infix with precedence looser than normal
function application.

Closes #512