Slight tweak monitor output for tty-multiprocess scenario
In particular, text at the very right-most edge can be obscured by the
cursor (depends on your style of cursor of course), so this commit
ensures that there is always a right-most space. This fixes the
previous situation for single processes (which output successive lines
of their output to the screen) when that didn't happen.

Hopefully finalising issue #526