mosml: Automatically move Holmake object files to sigobj
This simplifies handling of files that we want to use in core HOL
modules. Under Poly/ML, the same object files are already loaded into
the buildheap tool so the hacking of sigobj doesn't need to happen.

Along the way, make sure the regression test for issue #451 also works
properly under Moscow ML.