Fix bug where Holmake would die with Subscript exception
Issue was terminal environments where `stty size` and/or `tput
cols` (these are Unix shell commands) would return suggesting that the
width of the terminal window was 0. Fix the getWidth function to not
allow this.

Preserve a diagnostic message in poly/BuildCommand.sml that was of
some help in tracking this down.