Change big_record_size to be 100 by default
As discussed on the hol-info mailing list and in person on numerous
occasions, this technology is frequently extremely confusing. Users
should probably get around this problem themselves by defining
hierarchies of sub-records themselves.

A limit of 100 is not likely to be a practical issue in the near
future as records of this size will take painfully long to be defined,
and users will follow the manual's new advice to define
sub-records. (Just tested this: 10 fields: 0.08s; 20 fields: 1.0s; 30
fields: 6.2s; 50 fields: 89s.)