Fix Sub_and_cond.COND_ELIM_CONV when cond. exp is l.h.s. of arith op
The new implementation is now the first phase of
SUB_AND_COND_ELIM_CONV. The output is still not the same as the code
that ran in K10 but I'm not particularly concerned by this.  Indeed, I
don't know why this code is in the public API at all.

See (even-numbered) regression tests in selftest.sml for cases that
used to fail unhelpfully.

Closes #561