Report problems with simple definitions better
Primitive recursion and simple "abbreviation" definitions have always
allowed (redundant) universal quantification of the equations that the
user writes.  If they do this with a malformed definition though, the
Defn machinery catches the exception and tries to pass the equations
on to the more sophisticated machinery of wfrec definition. This then
causes confusing messages.

With regression tests

Thanks to Lorenz Leutgeb for the motivating example problems.