Further simplify Isabelle concurrency code port's CoreLib
The CoreLib's fold_map was semantically the same as our existing
foldl_map, just with arguments in different orders and with some
things curried instead of tupled.

Eventually, as much code as possible in this directory needs to move
out into our common shared library, leaving an unavoidable rump of
thread-code that is necessarily Poly/ML specific.