Delete some unreachable code
Cond_rewr.IMP_EQ_CANON is called on theorems that have already been
preprocessed by the same module's IMP_CANON.  The latter splits
conjunctions apart, deals with universal quantifiers and normalises on
the left of implications, so code for conjunctions and universals is
unnecessary in IMP_EQ_CANON.

The handling of Abbrev(f = \x. ...) assumptions does require the
elimination of some universal quantification, so this is done with a
call to SPEC_ALL before a recursive call to IMP_EQ_CANON as
previously.  (The recursive call is necessary to handle things like
rewrites underneath the Abbrev involving equalities on the left.  For
example:

   Abbrev (f = \x. x = y)

should simplify ``y = e`` to ``f e`` (yes, really), and

   Abbrev (X <=> (e1 = e2))

should simplify ``e2 = e1`` to ``X``.