Rewrite before_install.sh as install-poly.sh
Promote the contents of before_install.sh to be the de facto way of
setting up PolyML.

To make arguments for the script less confusing, use "T" and "F"
instead of empty/non-empty environment variables.
Also rename the environment arguments to be prefixed by POLY instead
of suffixed as I consider this more clear.

The renaming is in conformance with other {verb}-{noun} constructs in
/developers.