Fix bug in handling of bounded simp (e.g., with Once)
The bug was introduced by d71df467 and was apparent in a CakeML proof:
the commit made it impossible to use the simplifier to rewrite with
the Abbrev definition backwards (i.e., GSYM markerTheory.Abbrev_def)
because the variable on the LHS of that theorem has type bool. This is
a loopy rewrite, but should be permitted if the rewrite is bounded.

Equally, rewriting with the definition of I backwards would exhibit
the same problem, as long as one first instantiated the alpha in that
theorem to bool.