Add more utility functions to Portable
fold{l,r}2' are the obvious "primed" variants of ListPair.fold{l,r},
where priming means that the function takes its arguments curried, and
the accumulator is the last argument, as with foldl' and foldr'.

the_list and the_default handle option values and convert them to
something else: the_list to an empty or singleton list, and
the_default to the some value, or to the given default.