Add (commented-out) test case for bug to do with cache-fouling
Issue is that adding a theorem about constants to one's arith d.p.
context does not get seen as extending the context for the purposes of
the cache (because the theorem does not have any hypotheses), and so a
previous failure to prove the desired result, which did not have
access to the theorem, will cause subsequent attempts to also fail.

Progress with github issue #477