Fix bug in caching behind arith d.p's
It is important to track not only the hypotheses of the theorems in
the context (typically the assumptions in the goal), but also the
statements of the conclusions.  If for example, one adds a theorem to
the mix that doesn't have any hypotheses at all, the fact that this is
available should be seen as an extension of the context and the cache
shouldn't reject the new attempt to prove the goal.

Closes github issue #477