Move p/printing of set comprehensions out of src/parse
The code for this can be put into pred_set/src/pred_setpp, making the
core pretty-printer slightly simpler (which has to be a good thing).

Also fix the minor pretty-printing indentation bug identified in
github issue #610.

Making this a "user pretty printer" means that it takes precedence
over overloads. Introduce a new feature whereby adding a
(dummy) user-printer with the empty name causes terms matching this
pattern to be printed without any user-printers getting a chance to
fire. This gives terms that are handled by overloads a chance to be
seen by the core printing machinery. (User printers are checked for
first and can otherwise mask everything else.)

Closes #610