Fix Holmake's wildcard fn to interpret arg as multiple patterns
So, $(wildcard foo* bar*) now searches current directory for files
beginning with foo, and those beginning with bar. Previously, the
space would have been interpreted as a literal space. To get the
literal space, you need to escape it with a backslash (as per the 8th
testcase in the regression test directory).