Fix printing of embedded types
The symbol-merging avoidance technology was tripping up (part of the
issue is that the type pretty-printer's calls to add_string and
friends aren't mediated through the same interface as the term
p/printer's calls).

I'm far from sure as to why this wasn't an issue before. Anyway, it's
easy enough to tell the machinery that the "last string" printed was
whitespace (even though this isn't strictly speaking true), letting
the closing right-parenthesis be printed without an intervening space
being inserted.  (Certainly the type-parser should never have a
problem with a right-paren it's not expecting to see.)