Add a finally function to Portable
As with the finally clause in Java, this can be used to clean up after
a computation has completed, whether with an exception or not. As
Interrupts are caught so that the finisher still gets its chance, the
finisher shouldn't loop or take too long itself. (The user would have
to hit Interrupt twice.)