Reinstate Holmake's use of POLY_CLINE_OPTIONS
As per the Description manual, this gives the user a way of making
sub-processes (those that build theories, for example) pass options to
the Poly runtime when they are run.  Controlling the Holmake runtime
itself can be done simply by passing those options on the Holmake
command-line.

In other words

   Holmake --maxheap=8000

affects the Holmake process, while

   POLY_CLINE_OPTIONS = --maxheap=8000

affects the buildheap processes that turn script files into theories.