Move some theorems into pred_setTheory from CakeML
One, INFINITE_INJ_NOT_SURJ prompts a complete reworking of the very
old proof in this script that establishes something similar but only
for universal sets.  The new result is stronger, and is used to
establish the old one.

The new proof is in a similar style, but uses technology (METIS_TAC
and the simplifier) that wasn't available in HOL88.  It also uses a
newer result (the statement of the pigeonhole principle). It's rather
shorter therefore.