First steps towards parsing f ⦇ d1 ↦ r1; d2 ↦ r2 ⦈ for fn. UPDATE
The ASCII version of the same is f (| d1 |-> r1; d2 |-> r2 |) and of
course, the old (d1 =+ r1) ((d2 =+ r2) f) can also be used.

The pick of ⦇ ... ⦈ precludes using the same parentheses for
s-expressions in the example of the same name (hence the changes in
that example). The use of (| |) means that (||) no longer parses as
paren-escaping the || operator (hence the change to wordsScript.sml).

Pretty-printing still to be done.