Remove (| |) syntax in category theory example
Descendants of this theory might still use the Unicode version with
the lense parenthesis (⦇ ... ⦈), but the ASCII syntax was clashing
painfully with a custom syntax in this development that used vertical
bars, and was enclosed in parenthesis quite frequently.  The lexer
then saw the end of that syntax |) as one token, rather than two.