Fix minor problem caused by change to fn-update syntax
As of 782dc4800, the =+ identifier is no longer an overload for
UPDATE; it is instead an infix identifier that lives solely at the
concrete syntax level. After concrete syntax is removed (e.g., by
Absyn), the rest of the parsing process sees only the name UPDATE from
that point onwards. That makes a nonsense of $=+.  Instead, users
wanting to refer to the bare constant name need to write UPDATE.