Allow for string/char literals that aren't valid UTF8
UTF8 machinery was being applied to the contents of HOL's own string
literals, which led to problems if one wrote “"\172"” for example
because the SML level UTF8 machinery objected to \172. The fix is to
make sure that the SML/UTF8 stuff doesn't get called on string literal
tokens at all.

Closes #632