Make "quoted variables" ($var$(..)) syntax more robust
In particular, they may contain quotation symbols and
backslash-escaped characters that are not necessarily UTF8 safe.

Thus: “$var$(\172)” parses to a variable whose name (as revealed by
dest_var), is the string containing character 172. The pretty-printer
correctly reverses this.

There are yet two infelicities in all this:
- quoted variables with challenging contents (spaces, quotations) need
  to be preceded by whitespace to avoid completely confusing the lexer
- the pretty-printer doesn't safely print back quoted variables in
  such contexts, so you get something that won't parse

The latter is easy to fix.  The first is not.