Give standalone quotation filters option to control "interactivity"
Hitherto the form with no arguments

   unquote < foo > bar

was "interactive", and the form with two arguments

   unquote infile outfile

was not.  Now add a -i option for the second case to make it
interactive, and a -n option for the first to make it not interactive.
Also give a brief usage message for a standalone -h.

Interactivity controls whether or not certain special forms that
should only appear in script-files (and when the interactive system is
being used) get transformed. In particular, this controls the handling
of >- and the new Theorem <name> <quotation> <tactic> syntax.

The Poly/ML version of unquote is not used within the system but gets
the same changes to behaviour just to keep both versions equivalent.