Reorder examples that depend on finite_maps to follow src/finite_maps
Moving finite_mapTheory into src/finite_maps forced that directory to
come after src/n-bits (because patricia trees depend on the latter).
This in turn means that a couple of examples that use finite maps have
to come later too.

There is no change to the logical story here, just the order in which
example directories have to be built if they are to interleave with
the building of core HOL. (One could just wait for all of core HOL to
build before building examples and this problem wouldn't arise, but
occasional early failures in examples can be useful when testing for
regressions.)