Ensure -prefixed commands are allowed to fail with -jn for n > 1
The monitor function shouldn't be given responsibility for determining
whether or not a make-session should be aborted. Its role is really
just to show the user what is happening. Instead, alter the dep-graph
state that is consulted for new jobs to also include a "everything is
still ok" flag. This then gets updated to be false if a command fails
and keep_going is false, and if there is no - prefix.

Closes #655