Make the Poly/ML HOL REPL recover better after a Control-C
In particular, completely reset the lexing state. This allows one to
type an incomplete quotation (e.g., `foo), or an incomplete
string ("...), or an incomplete term/type (“...), hit Control-C and
then have the lexing state of the REPL return to the default,
top-level.

Hitting semi-colon <RETURN> is probably required to complete the
reset, but this is better than having to remember and then type a
delimiter form to match up against.