Don't suggest we can compile with Moscow ML 2.01
Now there is an immediate failure to configure if this is attempted.

We get into errors due to an inadequate basis implementation if we do.
This then allows simplification of the codebase, which tried to
provide a stripped down extension to the basis library.

Many options in Holmake get removed as a consequence.