Emacs mode fixes: standard hol command (M-h h) restarts in right dir
If a hol command happens in a buffer with an associated
directory (e.g., a script file), the hol process is started in the
same directory.  (Previously, a running HOL would just be left alone;
now it's explicitly killed and restarted in the right directory.)

Also fiddle with the ways in which the standard hol, hol-horizontal
and hol-vertical work in terms of trying to reuse existing HOL windows
rather than creating fresh ones.