Make Holmake c/line arg. "fooTheory" be treated as "fooTheory.dat"
This aims to make it easier for people to make specific theories
without needing to appreciate the fooTheory.{sig,sml,ui,uo,dat}
gawp. (Longer term, it would be great to hide/merge some of those
files entirely.)

Documented in the Holmake chapter of the DESCRIPTION manual, and
release notes. Tested in tools/Holmake/tests/theorytarget.