Change ThmSetData to store just theorem names in .dat files
The recent changes to this file caused theorem values to also be
stored; this is unnecessary (the references must be to existing
theorems), and wastes space in the .dat file, and presumably in what's
loaded.

I'm unsure as to why the stored name changed in the simpset
machinery (requiring a change to delsimps1Script.sml), but I'm going
to change the way simpset modification happens anyway.