Apply the quotation filter to all input files
Previously, a has_dq (“has double quotes”) function was applied to
test files to see if they should be filtered. With Theorem syntax used
in files that didn't include any Unicode or backquotes, this analysis
was wrong. Nor is there any need for it.  It's simpler to just apply
it to everything.