Revise the way "exclusion/deletion" strings are handled by simpLib
The srw_ss() simpset includes the list.APPEND_ASSOC rewrite, and this
new method allows for the strings "APPEND_ASSOC" and
"list.APPEND_ASSOC" to remove that rewrite. Importantly, the former
will hit other theorems of that name from other theories (though I
admit that's yet to be part of a regression test).