Make hol-horizontal (M-h 3 or M-h H w/horizontal) have 80 col window
If the original window (the one being split) is too narrow, giving 80
columns to the new one is obviously regarded as impossible, and the
split then appears to be 50/50. But, by observation, if the original
is even 110 columns wide, then the new HOL windows does get to achieve
its width of 80 (making for a very shrunken original).