Prove a version of EXISTS_UNIQUE_ALT in boolScript.sml
There is an EXISTS_UNIQUE_ALT proved in boolLib; this moves a version
to the DB (match and find will see it) called EXISTS_UNIQUE_ALT'. It
flips the direction of the equality on the RHS, so that rewriting a
?!x. P x in the assumptions will give a new assumption (after
stripping) that looks like

  !y. P y <=> (y = x)

My theory is that this direction (y = x) will be more useful as an
elimination.  (And if not, use the original EXISTS_UNIQUE_ALT.)