Update comment on DPLL function's speed in TUTORIAL's tools chapter
On my 4 year old Macbook Pro it now happens in less than a second.
Clearly I need to be able to feed the output of the SML time call into
the LaTeX sources. (It wouldn't be too hard to have the SML code check
that the timing *was* less than a second, and abort if not...)